creation of a navigation between pages gui contains the mainframe and the navigation gui_rumor contains the rumor page gui_store contains the store page page is a parent of every pages to allow the superposition of pages and is used by in GUI.py to elevate the correct page and display it above every other pages
