+++ src/page.py
@@ -1,0 +1,9 @@
+import tkinter as tk
+
+
+class Page(tk.Frame):
+    def __init__(self, *args, **kwargs):
+        tk.Frame.__init__(self, *args, **kwargs)
+
+    def show(self):
+        self.lift()
